Avoid messing with the group-leader and user process in test-shell
Summary:
# Context
The test-shell runs the test on a separate node. In this node, the group-leader of the main process is replaced by the group leader of the main node and the user process by one that forwards messages to that GL. This is so that stdout is thisplayed on the shell of the main node. Notice that this forces both nodes to have distribution enabled.

To start the test-runner node, the test-shell uses the same logic as we use for the CT runner. And there, we want to move to a model where the stdout of the runner is forwarded to the main one via the port.

# This diff
As a first step, update the test-shell so that it gets stdout/stderr from the port. For this we start using a previously unused `ReplayIO` option, so that the test-shell and the CT runner can have different behaviour for now.

Reviewed By: TheGeorge

Differential Revision: D87566912

fbshipit-source-id: 1dfa6a72bfc648e4f564a78a7a7a37d3b312ad3a