Avoid long stdout lines being truncated in Erlang tests
Summary:
# Context
We recently change the way Erlang tests run, so that the test-executor node running the CT suite has its stdout piped to the test-binary node, who logs for later processing

The port opened on the test-binary is using the `{line, N.}` option, which means that lines of size at most `N` are sent in a single `{eol, L}` message, but longer lines are split in one or more `{noeol, Ln}`, and a final `{eol, L}`.

# Problem
The logic handling messages from the port is only handling `{eol, L}` messages, so we'd be missing the prefix of very long lines.

In practice, N = 1Kb, so in practice this had little impact

# This diff
Handle the `noeol` message as well

Reviewed By: michalmuskala

Differential Revision: D88007158

fbshipit-source-id: 1584edbd9bbd9b374cd3f3f6cbf2c8b4d3d3ef5c