Show stdout while debugging CT tests
Summary:
# Context
In the last diffs we did a major refactoring in the way we handle stdout while running Erlang tests. We finally come out of that rabbit hole, as we have now everything in place to solve the main issue we were trying to fix.

# Problem
At the moment, all stdout from the test-executor node running the CT suite is saved to a file and later processed to attach the relevant stdout to each test-case.

However, when debugging a test with edb, this means that the user can't see the stdout as the execution of the test progresses.

# This diff
We extend the ct_stdout processing with a flag that, when enabled, will not only log each line of stdout to a file, but it will output it to the test-binary's stdout as well.

We then configure things so that this flag is turned on whenever the edb setup code is to be injected in the test-executor node.

Reviewed By: TheGeorge

Differential Revision: D88013185

fbshipit-source-id: a011454d6aad174cbf06dd8fdfc944d2b6f924c5