Remove progress markers from stdout on the fly
Summary:
# Context
See D88003880.

# Problem
We have now all the infrastructure in place to collect and filter-out stdout progress markets as it is streamed from the test-executor to the test-binary node, but we are still instead writing the raw stdout to a file and postprocessing it

# This diff
We make it so that the process handling the port to the test-executor now processes stdout as it comes. The "raw" stdout file no longer exists, the one we write is already filtered out.

Reviewed By: TheGeorge

Differential Revision: D88012398

fbshipit-source-id: f46df71132e0918433b400dd2d8090cc23abd8c0