Make test progress markers identifiable
Summary:
# Context
We now have made it so that when running Erlang tests, stdout from the "test-executor" node running the tests is piped to the "test-binary" supervising the process.

The test-binary currently just logs the piped stdout to a file, and later rereads this file to identify and remove the "progress markers" that delimit the portion of stdout that correspond to each testcase, while also collecting these bits of stdout.

We are almost ready to make it so that stdout is also emitted on the test-binary stdout while debugging.

# Problem
However, if we were too enable this now, then the progress markers would also be emitted to the user. This will pollute the output and is an implementation detail that shouldn't be leak to the user.

# Idea
What we would like is to be able to identify and remove the progress markers before logging the received stdout to the file. This means that we need a way to "identify" that a line is a progress marker as we receive it. This is in contrast to the current approach where we get from the test-results the progress-markers emitted, so we only need to match against them.

# This diff
Here we introduce a "fingerprint" that we will use to identify that a line is a progress-marker. We create a random fingerprint on each suite execution, so we don't get false positives.

Then we change the current logic to extract the stdout of each test, so that we depend only on the fingerprint and not in knowing the actual markers emitted.

In the next we will move this processing to the port that reads stdout.

Reviewed By: TheGeorge

Differential Revision: D88003880

fbshipit-source-id: 997643453348b2050ee9c5adf0c77046a9f7bb42