git_fetch: Allow specification of the used object format
Summary:
This PR adds a new (optional) attribute `object_format` to the git_fetch rule. If given, the values of the attribute must be one of `sha1` or `sha256`, and its value is used for the initialization of the underlying Git repository (e.g. `git init --object-format sha1`).

The given `rev` is validated according to the chosen object format; If none is provided we still check if the `rev` is a valid SHA1 digest, but we do not pass the `--object-format` option to Git.

X-link: https://github.com/facebook/buck2/pull/1111

Reviewed By: ahornby

Differential Revision: D83531241

Pulled By: dtolnay

fbshipit-source-id: d5539944e499938b6efd81bb075a6a3569ab9aef