new riscv64imacfv toochain
Summary:
* adds riscv64imacfv toolschain
* enables that toolchain for Athena/Iris rust PE targets (PESTr tests)

Add a new tooolchain, same as the existing riscv64imacf, but with a "v". This allows the use if risc-v vector instructions.

bcfg file was generated by:
  fbpython "$(hg root)/tools/buckconfigs/fbsource/generated/generate.py"

-------------
**Artemis & vectors**
Once +v is enabled the compiler automatically vectorizes code.  There is
apparently no way to completely turn this off.  Since Artemis PE's have two
different RISC-V CPU's, one of which does not support vector instructions,
this toolchain can't be used on Artemis.

I looked for a master switch to stop any vectorization instructions from being added to the binary but found nothing.  I did find some related config options, which I tried passing in as "llvm-args":
```lang=text
        "-riscv-v-vector-bits-min=0",
        "-vectorize-loops=false",
        "-vectorize-slp=false",
        "-force-vector-width=0",
        "-scalable-vectorization=off",
        "-riscv-disable-vector-mask-mutation",
        "-riscv-v-fixed-length-vector-lmul-max=0",
        "-enable-epilogue-vectorization=false",
```
and, I found sometimes these could reduce the number of vector instructions in the binary but not eliminate them.  For instance `-riscv-v-vector-bits-min=0"` reduced the number of instructions starting with the letter v from 900 to 300.

As a result trying to use the vector enabled toolchain on Artemis does not work.  Even if the code returns as soon as it can determine if it is core-A or not, core-A will try to execute a vector instruction before it reaches that point and crash as a result.

Since Athena+ PE's have vector instruction support on both cores this toolchain works on those targets.

Reviewed By: dtolnay

Differential Revision: D87782278

fbshipit-source-id: 9d546f264032c7bf500b17a1439e4a7277071be5