Support content based actions in archiving
Summary:
Adding support for content based paths to archive actions. Depends on if the cxx toolchain supports content based actions, same as other cxx actions.

Reland of D86930889

Disabling haskell archiving from using content based paths, because it seems that toolchain breaks when a symlink is added in, and it reuses the same archive action creation as cxx.

Reviewed By: IanChilds

Differential Revision: D87938880

fbshipit-source-id: 0a92b59cf1bdf97bc0a428ca52f710509029723c