No --file-prefix-map for incremental remote
Summary:
Swift Incremental can't sanely use `--file-prefix-map` for non-local compile actions, because it results in changing arguments from run to run, as the path can change from machine to machine.

This diff turns off `--file-prefix-map` for Swift Incremental when file-hashing is enabled.

The knock-on effects of this are that absolute paths are embedded in DWARF data for compiler-generated Swift code (i.e. Swift Macro generation). To avoid this issue, we ensure that DWARF version 5 is used, which embeds generated source inline in the data (without a path to a generated file). Note that we are already using DWARF 5 in anything that targets iOS18+.

Reviewed By: rmaz

Differential Revision: D87788794

fbshipit-source-id: e407ca5516090712a9c2666732f15d0387149cd9