Track emit-module.dia file
Summary:
# Context
There are certain files that are generated during a Swift compile that are not currently tracked by Buck2. If we do not track them, they won't be available as incremental inputs in RE actions, for instance. So for each, we either need to
1) Not emit the files (sometimes safe, as if they were important output for another action, they would be tracked).
2) Track the files

# This diff

The .dia files are all currently tracked as Artifacts, except for one - the swiftmodule's artifact, which defaults to "${MODULE_NAME}.emit-module.dia".

This diff adds an artifact to track this for incremental compiles when `supports_serialized_errors` is enabled.

# Why not set this alongside where the the swiftmodule output is enabled?

Ideally this would be added alongside an entry for the swiftmodule, however the swiftmodule path is set by a command-line arg, and it can't currently be moved to the  output-file-map as Swift incremental mode does not respect it.

Reviewed By: rmaz

Differential Revision: D87788796

fbshipit-source-id: 51b5afdb2b2d6dfb7be23f54105487a0baf56286