fix scrubber logic for content based paths
Summary:
Focused debugging was not working with content based path turned on.

Context post: https://fb.workplace.com/groups/410668900045165/posts/1489582175487160

The cause seems to be in the scrubbing logic, where it was making assumptions regarding the output path for each artifact.

More context is in the newly added comments, which shows that with content based path turned on, the path is now different.

The fix for this problem needs to consider two factors:
- Should work for both content based path turned on and off. (Since not all apps have content based path turned on right now)
- It should be fast, since the scrubber is being executed on thousands of artifacts during each build.

With this in mind, I think that having some complicated logic that checks whether a path is content based adds too much overhead and can cause some performance issues. The proposed solution is a bit "hacky" in terms it will return two "trimmed" path, considering both cases. And we will only remove this artifact only if both "trimmed" path are not in `focused_targets_output_paths`.

The downside of this solution is that it "might" include more paths than expected in the output,=

Reviewed By: milend

Differential Revision: D88386746

fbshipit-source-id: 412169a8c90e565befd9d25134ff6e937ee8d83b