Fix simulator runtime selection for 26.0.1
Summary:
Investigating the problem in [this post](https://fb.workplace.com/groups/ios.testing/permalink/30096535963301632/), I discovered that `simctl` does not match runtimes by their *name*, it actually matches them by their *platform + runtime version*.

This is important because a runtime called 'iOS 26.0' with the version '26.0.1' will not be matched to the string 'iOS 26.0' when `simctl create` is called. The string 'iOS 26.0.1' has to be used to specify the runtime. (It's unclear if this is a bug in `simctl`, but this behavior goes against `simctl`'s own help text.)

This diff updates the resource broker to pass the platform + runtime version for the os version instead of the runtime name. This fixes local testing simulator creation when a runtime with a patch version is installed on the user's machine.

Reviewed By: milend

Differential Revision: D88602561

fbshipit-source-id: 3603242ca84f96b67569ee6278dc39980edda918