Don't try and emit 'emit-module.dia' for Xbat
Summary:
To support Swift Incremental with remote artifacts and file hashing, we need to ensure the `${modulename}.emit-module.dia` diagnostics file is declared as an Artifact if it is being emitted.

It is only emitted when the right flags to generated it are passed to the compiler (via `add_serialized_diagnostics_output`).

Previously we have assumed that it is always emitted for incremental builds and the logic was in `swift_incremental_support.bzl`. However this is not correct as Xbat builds do not include this output. In this case it is wrong to add this artifact as it will never be generated, even if we put it as an entry in the file-output-map.

This diff moves the logic for adding this to the file-output-map and tracking the artifact to inside the call to `add_serialized_diagnostics_output`.

Reviewed By: rmaz

Differential Revision: D88748466

fbshipit-source-id: ae2588f74d24e64f421c96c2b7cec988f7a4ce9b