Add resource broker replays for debugging
Summary:
Similar to D88276740 (which added debug replays for `xctest_runner`), this diff adds debug replays for the iOS `resource_broker`. By adding the `DUMP_REPLAY` env to a test command, the `resource_broker` will create a script to replay the `resource_broker` command that was used.

The replay script is written to: `/tmp/resource_broker_replay.sh`

*Note: The `DUMP_REPLAY` env is not currently passed through `tpx` to the `resource_broker`, so there will be a follow up diff to fix that.*

Reviewed By: patskovn

Differential Revision: D88895174

fbshipit-source-id: 81d93eb75e1b7b48ecf917ebfa536dc05c3feb7b