swap `selects.apply` over to the native `select_map(..., recurse=True)`
Summary:
Note that this is a re-do of D86416543 now that D87708260 managed to land and update the dotslashed autodeps executable.

Buck bumped to include the new `select_map` on Nov 20 in D87543825.

When `recurse=True` becomes the default, we can codemod `selects.apply(...) => select_map(...)`. But existing uses need to remain until we either DO that, or we codemode `selects.apply(...) => select_map(..., recurse=True)`. Work tracked in T245559941

Reviewed By: JakobDegen

Differential Revision: D88951618

fbshipit-source-id: 057e2d5d707caf679ee8ba1df83512e2abe1d98d