S585110: supply both *.o and *.dwo files for `bitcode` lto objects to bolt
Summary:
Under some (unknown) conditions, clang will NOT create a *.dwo file, and instead will leave the `.debug*.dwo` sections in the *.o file. This is at LEAST true when ThinLTO is in use, AND the object is a "bitcode" object, even if you pass the `dynamic_optimize()` action `--split-dwarf=split`. For these, the `external_debug_info` needs to include the *.o file, too.

This may not be the ONLY problem, either... bugs in either llvm's creation of the debug info in thinlto mode or in bolt's processing of llvm's output may also be contributing. Actually having all the debug info present during the bolt step should be a prerequisite for further investigation.

Reviewed By: akatrevorjay

Differential Revision: D89057752

fbshipit-source-id: f1ccd368942055856172066a95ea640a32b244dc