Forward providers in apple_finalize_bundle
Summary:
`apple_finalize_bundle` is used for MacOSX apps to codesign and return a binary that can be packaged.

Recently, in the `apple_package` rule, we began to return a provider that contained certain information (linker maps, dsyms, etc) which do not get passed on from `apple_finalize_bundle`. This makes it impossible to get this information for these apps.

Forward these providers and undo the workaround fix in D89151848.

Reviewed By: cosminconst

Differential Revision: D89224481

fbshipit-source-id: 80d68694c11f4a012aa5f82ed6be2ef015ae7a2d