Detect duplicate link labels
Summary:
The graph data structures in the LinkableGraph provider use `Label` to identify dependency edges in the linkable graph.

https://www.internalfb.com/code/fbsource/[8d137950db036a493826ac0540b37f263ce5d6e1]/fbcode/buck2/prelude/linking/linkable_graph.bzl?lines=66%2C72-73%2C120

This makes it nonsensical to depend on multiple distinct linkable nodes having the same label. A given Label (which includes target platform configuration) should only ever be associated with a single linkable node; and same for other things like external debuginfo.

This diff adds an assertion failure if we ever observe different data with the same label in the Rust rules. This eliminates edge cases that we would otherwise need to consider when translating more of the linkable graph processing to transitive sets in upcoming diffs.

Reviewed By: IanChilds

Differential Revision: D89127275

fbshipit-source-id: f7da51603bc043b018fca40b75c90180f43c7a0c