Improve robustness
Summary:
This android automation is extremely flaky on horizon os emulator. This diff adds some known tricks used on test automation to prevent spurious failures during `reboot` and `root`.

Check this test results for an example of failures in the horizon os emulator:

https://www.internalfb.com/intern/test/844425208253572?ref_report_id=0

Example of institutional knowledge:

https://fb.workplace.com/groups/418844805594414/permalink/1525578014921082/

Also, we should take some inspiration of Gauntlet Internals on how to setup the Horizon OS device to avoid this type of issues:

Differential Revision: D89589429

fbshipit-source-id: 26a99cee8e2c8ac91a91d8e64c0f58b05ab8aabc