Bump 0.7.0 build with rebase (torch 2.10)
Summary:
Rebase complete.

This is a port of stable fairseq2 v0.7.0 to fbcode. Additionally, some logic has been added to bind to sentencepiece headers in v0.6.0 so that large files do not need to be reuploaded.

PyTorch version used for build has been updated to 2.10 for both 0.6 and 0.7. The buckify patch has been updated to reflect this. (PyTorch 2.10 has been tested on CPU and GPU from FBCode).

Reviewed By: jermenkoo

Differential Revision: D89501661

fbshipit-source-id: 92539debcb0a1af348e82fc91d0b353d2ae51920